Cleaner interface in api.jl
Implemented generate_path_from_config_and_state,
and produced closures in NamerInterface ->
generate_path
and
generate_path_with_previous_date
Both return a file_path, and generate the folders in between.
The latter reuses the cached date, to make it easy to produce a
collection of files with the same timestamp but different stems,
for when a collection is made up of multiple files, images, and meta
data.

Collection folder incrementing behavior not yet implemented.

Testing with scripts/run-example.sh
and api.jl -> api_demo()
to test operation.